requirements.txt
snowflake-connector-python[pandas]
nba_api